Add files via upload
This file tracks the pre-processing done in Excel to convert the OFSP_original file to the OFSP_english file and finally to the OFSP_rpts_rmv file. The OFSP_rpt_rmv file is in English and repeated columns have been removed. For repeated column headings, the column containing the most data is retained.